+++ app/__init__.py
@@ -26,19 +26,27 @@ def create_app():
 
     # Register blueprints
     from app.views.auth import auth_bp
-
     app.register_blueprint(auth_bp)
 
     from app.views.dashboard import dashboard_bp
-
     app.register_blueprint(dashboard_bp)
+    
+    from app.views.you_page import you_page_bp
+    app.register_blueprint(you_page_bp)
+    
+    from app.views.matches_page import matches_page_bp
+    app.register_blueprint(matches_page_bp)
+    
+    from app.views.about_us_page import about_us_page_bp
+    app.register_blueprint(about_us_page_bp)
+    
+    from app.views.find_page import find_page_bp
+    app.register_blueprint(find_page_bp)
 
     from app.views.profile_manager import profile_manager_bp
-
     app.register_blueprint(profile_manager_bp)
 
     from app.views.profile import profile_bp
-
     app.register_blueprint(profile_bp)
 
     return app
