--- app/extensions/extensions.py
@@ -1,4 +1,0 @@
-from flask_login import LoginManager
-
-login_manager = LoginManager()
-
