+++ app/models/__init__.py
@@ -1,0 +1,14 @@
+from .models import (
+    User,
+    db,
+    MatchingPreferences,
+    Profile,
+    Gender,
+    Interests,
+    Photo,
+    Profile,
+    profile_interests_table,
+    LikingAssociation,
+    BlockingAssociation,
+    DateProposal,
+)
