+++ app/models/models.py
@@ -8,16 +8,18 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy import Column, DateTime, Enum, ForeignKey, Integer, String, Table
 from sqlalchemy.ext.associationproxy import association_proxy
 from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import Mapped, mapped_column, mapper, relationship
+from sqlalchemy.orm import Mapped, mapped_column, relationship
 from werkzeug.security import check_password_hash, generate_password_hash
 
 db = SQLAlchemy()
 Base = declarative_base()
 
-available_genders = ["male", "female", "non_binary", "other"]
-available_genders_display = ["Male", "Female", "Non binary", "Other"]
 
-Gender = Literal[*available_genders]
+class Gender(enum.Enum):
+    MALE = "Male"
+    FEMALE = "Female"
+    NON_BINARY = "Non binary"
+    OTHER = "Other"
 
 
 class Interests(db.Model):
@@ -76,14 +78,10 @@ class User(db.Model, UserMixin):
         back_populates="blocking",
     )
     sent_proposals: Mapped[list["DateProposal"]] = relationship(
-        "DateProposal",
-        back_populates="proposer",
-        foreign_keys="DateProposal.proposer_id",
+        "DateProposal", back_populates="proposer", foreign_keys="DateProposal.proposer_id"
     )
     received_proposals: Mapped[list["DateProposal"]] = relationship(
-        "DateProposal",
-        back_populates="recipient",
-        foreign_keys="DateProposal.recipient_id",
+        "DateProposal", back_populates="recipient", foreign_keys="DateProposal.recipient_id"
     )
 
     def set_password(self, password):
@@ -99,15 +97,7 @@ class Profile(db.Model):
     user_id: Mapped[int] = mapped_column(ForeignKey("user.id"), unique=True)
     user: Mapped["User"] = relationship("User", back_populates="profile")
     name: Mapped[str] = mapped_column(String(100), nullable=False)
-    gender: Mapped[Gender] = mapped_column(
-        Enum(
-            *get_args(Gender),
-            name="gender",
-            create_constraint=True,
-            validate_strings=True,
-        ),
-        nullable=False,
-    )
+    gender: Mapped[Gender] = mapped_column(Enum(Gender), nullable=False)
     year_of_birth: Mapped[int] = mapped_column(Integer, nullable=False)
     description: Mapped[str] = mapped_column(String(500), nullable=True)
     photo_id: Mapped[Optional[int]] = mapped_column(ForeignKey("photo.id"))
@@ -162,19 +152,10 @@ class BlockingAssociation(db.Model):
 class UserGenderPreference(db.Model):
     __tablename__ = "user_gender_preferences"
     matching_preferences_id: Mapped[int] = mapped_column(
-        ForeignKey("matching_preferences.id", ondelete="CASCADE", onupdate="CASCADE"),
-        primary_key=True,
+        ForeignKey("matching_preferences.id", ondelete="CASCADE", onupdate="CASCADE"), primary_key=True
     )
     matching_preferences: Mapped["MatchingPreferences"] = relationship("MatchingPreferences", back_populates="genders")
-    gender: Mapped[Gender] = mapped_column(
-        Enum(
-            *get_args(Gender),
-            name="gender",
-            create_constraint=True,
-            validate_strings=True,
-        ),
-        primary_key=True,
-    )
+    gender: Mapped[Gender] = mapped_column(Enum(Gender), primary_key=True)
 
 
 class MatchingPreferences(db.Model):
@@ -182,16 +163,9 @@ class MatchingPreferences(db.Model):
     id: Mapped[int] = mapped_column(primary_key=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("user.id"), unique=True)
     user: Mapped["User"] = relationship("User", back_populates="matching_preferences")
-    genders = relationship(
-        "UserGenderPreference",
-        back_populates="matching_preferences",
-        cascade="all, delete-orphan",
-    )
+    genders = relationship("UserGenderPreference", back_populates="matching_preferences", cascade="all, delete-orphan")
     gender_preferences = association_proxy(
-        "genders",
-        "gender",
-        creator=lambda gender: UserGenderPreference(gender=gender),
-        cascade_scalar_deletes=True,
+        "genders", "gender", creator=lambda gender: UserGenderPreference(gender=gender), cascade_scalar_deletes=True
     )
     lower_difference = mapped_column(Integer, nullable=False)
     upper_difference = mapped_column(Integer, nullable=False)
